+++ axum-login/tests/integration-test.rs
@@ -1,3 +1,5 @@
+#![cfg(feature = "macros-middleware")]
+
 use std::{
     collections::HashMap,
     process::{Child, Command},